Support new features in template-haskell-2.15.*
This adds support for `AppKindT` (visible kind applications) and
`ImplicitParamT` (implicit parameter constraints) in `expandSyns`.
I had to do some slight refactoring in `expandSyns'` to make this
happen since it is possible for visible kind arguments and other
arguments to be in any order (e.g., `Foo @T1 T2 @T3 T4`), so I
needed to add an auxiliary `TypeArg` data type to distinguish
between them.

This builds on top of the machinery in #14, so this subsumes #14.
Also subsumes #12.