Make th-expand-syns a shim on top of th-abstraction (#24)
This replaces `th-expand-syns`' substitution and type expansion machinery
with that of `th-abstraction`, which provides essentially the same
functionality. After this change, `th-expand-syns` becomes, for the most
part, a small shim on top of `th-abstraction`. There are still a few
pieces of functionality which `th-expand-syns` provides over `th-abstraction`
(see the `README`), but these are mostly minor things.

Fixes #23.